pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mms-survey"
version = "0.2.0"
description = "Contains Python code for processing MMS survey data"
license = "GPL-3.0-only"
authors = ["Tien Vo <tien.a.vo@proton.me>"]
readme = "README.md"
packages = [{include = "mms_survey", from="src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
xarray = "^2023.8.0"
zarr = "^2.16.1"
tqdm = "^4.66.1"
astropy = "^5.3.2"
psutil = "^5.9.5"
pathos = "^0.3.1"
cdflib = "^1.1.0"
dask = "^2023.8.1"
tvolib = "^0.5.1"
numpy-quaternion = "^2022.4.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
jupyterlab = "^4.0.7"
jupyterlab-vim = "^4.1.0"
jupyterlab-widgets = "^3.0.9"
ipywidgets = "^8.1.1"

[tool.black]
line-length = 79
target-versions = ["py311"]

[tool.isort]
profile = "black"
